+++ vue/components/ui/atoms/textarea/textarea.vue
@@ -62,7 +62,7 @@
     font-family: $font-family;
 }
 
-.textarea:hover:not(:disabled) {
+.textarea:hover:not(:disabled):not(:focus) {
     border-color: $aqcua-blue;
 }